+++ MercadoPagoSDK/MercadoPagoSDK/Flows/InitFlow/InitFlow+Services.swift
@@ -30,7 +30,7 @@ extension InitFlow {
     }
 
     func validatePreference() {
-        let errorMessage = model.properties.checkoutPreference.validate()
+        let errorMessage = model.properties.checkoutPreference.validate(privateKey: model.properties.privateKey)
         if errorMessage != nil {
             let customError = InitFlowError(errorStep: .ACTION_VALIDATE_PREFERENCE, shouldRetry: false, requestOrigin: nil, apiException: nil)
             model.setError(error: customError)
@@ -83,6 +83,10 @@ extension InitFlow {
             defaultInstallments = String(dInstallments)
         }
 
+        var maxInstallments: String?
+        let mInstallments = model.properties.checkoutPreference.getMaxAcceptedInstallments()
+        maxInstallments = String(mInstallments)
+
         let hasPaymentProcessor: Bool = model.properties.paymentPlugin != nil ? true : false
         let discountParamsConfiguration = model.properties.advancedConfig.discountParamsConfiguration
         let marketplace = model.amountHelper.preference.marketplace
@@ -92,7 +96,7 @@ extension InitFlow {
         //payment method search service should be performed using the processing modes designated by the preference object
         let pref = model.properties.checkoutPreference
         serviceAdapter.update(processingModes: pref.processingModes, branchId: pref.branchId)
-        serviceAdapter.getPaymentMethodSearch(amount: model.amountHelper.amountToPay, exclusions: exclusions, oneTapInfo: oneTapInfo, payer: model.properties.paymentData.payer ?? PXPayer(email: ""), site: SiteManager.shared.getSiteId(), extraParams: (defaultPaymentMethod: model.getDefaultPaymentMethodId(), differentialPricingId: differentialPricingString, defaultInstallments: defaultInstallments, expressEnabled: model.properties.advancedConfig.expressEnabled, hasPaymentProcessor: hasPaymentProcessor, splitEnabled: splitEnabled), discountParamsConfiguration: discountParamsConfiguration, marketplace: marketplace, charges: self.model.amountHelper.chargeRules, callback: { [weak self] (paymentMethodSearch) in
+        serviceAdapter.getPaymentMethodSearch(amount: model.amountHelper.amountToPay, exclusions: exclusions, oneTapInfo: oneTapInfo, payer: model.properties.paymentData.payer ?? PXPayer(email: ""), site: SiteManager.shared.getSiteId(), extraParams: (defaultPaymentMethod: model.getDefaultPaymentMethodId(), differentialPricingId: differentialPricingString, defaultInstallments: defaultInstallments, expressEnabled: model.properties.advancedConfig.expressEnabled, hasPaymentProcessor: hasPaymentProcessor, splitEnabled: splitEnabled, maxInstallments: maxInstallments), discountParamsConfiguration: discountParamsConfiguration, marketplace: marketplace, charges: self.model.amountHelper.chargeRules, callback: { [weak self] (paymentMethodSearch) in
 
             guard let strongSelf = self else {
                 return