+++ MercadoPagoSDK/MercadoPagoSDK/Services/Models/PXCheckoutPreference.swift
@@ -76,6 +76,7 @@ import Foundation
         }
     }
 
+    open var backUrls: PXBackUrls?
     internal var binaryModeEnabled: Bool = false
     internal var pxAdditionalInfo: PXAdditionalInfo?
 
@@ -138,6 +139,7 @@ import Foundation
         case site
         case marketplace
         case additionalInfo = "additional_info"
+        case backUrls = "back_urls"
         case branchId = "branch_id"
         case processingModes = "processing_modes"
     }
@@ -159,6 +161,7 @@ import Foundation
         self.init(id: id, items: items, payer: payer, paymentPreference: paymentPreference, siteId: siteId, expirationDateTo: expirationDateTo, expirationDateFrom: expirationDateFrom, site: site, differentialPricing: differentialPricing, marketplace: marketplace, branchId: branchId, processingModes: processingModes)
         self.additionalInfo = try container.decodeIfPresent(String.self, forKey: .additionalInfo)
         populateAdditionalInfoModel()
+        self.backUrls = try container.decodeIfPresent(PXBackUrls.self, forKey: .backUrls)
     }
 
     /// :nodoc:
@@ -173,6 +176,7 @@ import Foundation
         try container.encodeIfPresent(self.differentialPricing, forKey: .differentialPricing)
         try container.encodeIfPresent(self.marketplace, forKey: .marketplace)
         try container.encodeIfPresent(self.additionalInfo, forKey: .additionalInfo)
+        try container.encodeIfPresent(self.backUrls, forKey: .backUrls)
         try container.encodeIfPresent(self.branchId, forKey: .branchId)
         try container.encodeIfPresent(self.processingModes, forKey: .processingModes)
     }